[BUGFIX] Skip empty HTML body in Fluid mails
TYPO3 v11.5.27 introduced a change in the way FluidEmails are processed
which results in the HTML body of Fluid mails to sometimes be empty.

Avoid CSS inline style conversion in this case since this would fail
with an error and is pointless to begin with.

See https://review.typo3.org/c/Packages/TYPO3.CMS/+/78251